split input
